Fix モーダル閉じたらプレビューもリセット
